+++ flutter_modular/lib/src/core/models/child_route.dart
@@ -12,10 +12,21 @@ class ChildRoute<T> extends ModularRouteImpl<T> {
     List<ModularRoute> children = const [],
     required Widget Function(BuildContext, ModularArguments) child,
     List<RouteGuard>? guards,
+    String? guardedRoute,
     TransitionType transition = TransitionType.defaultTransition,
     CustomTransition? customTransition,
     RouteBuilder<T>? routeGenerator,
     Duration duration = const Duration(milliseconds: 300),
-  }) : super(routerName,
-            routerOutlet: [], duration: duration, child: child, routeGenerator: routeGenerator, customTransition: customTransition, children: children, guards: guards, transition: transition);
+  }) : super(
+          routerName,
+          routerOutlet: [],
+          duration: duration,
+          child: child,
+          guardedRoute: guardedRoute,
+          routeGenerator: routeGenerator,
+          customTransition: customTransition,
+          children: children,
+          guards: guards,
+          transition: transition,
+        );
 }